refactor: rewrite auth
