+++ src/context/index.js
@@ -1,7 +1,5 @@
 import React from 'react';
 
-const AuthContext = React.createContext({
-  isLoggedIn: false,
-});
+const AuthContext = React.createContext(null);
 
 export default AuthContext;
